Overwrite `apache_vhosts` with modified verisons for Let's Encrypt
The addition of the Apache redirect for missing files (48c6fe5) caused
a regression with Let's Encrypt:

Because the tasks for the redirect call the `geerlingguy.apache` role
to modify the vhosts file, it was overwriting the changes made for
Let's Encrypt.

This fixes the problem by updating `apache_vhosts` and
`apache_vhosts_ssl` with the rebuilt versions for Let's Encrypt.