Explicitly typecast `drupal_update_security_only`
Per Ansible documentation, command line extra vars are interpreted
as strings:

"Values passed in using the key=value syntax are interpreted as strings.
Use the JSON format if you need to pass in anything that shouldn’t be a
string (Booleans, integers, floats, lists etc)."

That causes incorrect interpretation when using
`-e "drupal_update_security_only=False"` to run all updates.  It
interprets the `False` as a string ("False"), which is always converted
to `(bool) True` due to Python's truthiness rules.

This change adds explicit typecasting so that the above will actually
be interpreted `(bool) False` and disable the security-only behavior.